+++ lib/src/xml/extensions/string.dart
@@ -1,12 +1,27 @@
 import '../exceptions/type_exception.dart';
+import '../nodes/cdata.dart';
 import '../nodes/document_fragment.dart';
 import '../nodes/node.dart';
-import 'has_children.dart';
-import 'has_parent.dart';
-import 'has_writer.dart';
+import '../nodes/text.dart';
+import 'descendants.dart';
+import 'mutator.dart';
 
-/// Mixin for nodes with XML contents.
-mixin XmlHasXml implements XmlChildrenBase, XmlParentBase, XmlHasWriter {
+extension XmlStringExtension on XmlNode {
+  /// Return the concatenated text value its descendants.
+  String get innerText => descendants
+      .where((node) => node is XmlText || node is XmlCDATA)
+      .map((node) => node.value)
+      .join();
+
+  /// Replaces the children of this node with text contents.
+  set innerText(String value) {
+    XmlNodeTypeException.checkHasChildren(this);
+    children.clear();
+    if (value.isNotEmpty) {
+      children.add(XmlText(value));
+    }
+  }
+
   /// Return the markup representing this node and all its child nodes.
   String get outerXml => toXmlString();
 
@@ -18,7 +33,7 @@ mixin XmlHasXml implements XmlChildrenBase, XmlParentBase, XmlHasWriter {
 
   /// Replaces the markup representing the child nodes of this node.
   set innerXml(String value) {
-    XmlNodeTypeException.checkHasChildren(this as XmlNode);
+    XmlNodeTypeException.checkHasChildren(this);
     children.clear();
     if (value.isNotEmpty) {
       children.add(XmlDocumentFragment.parse(value));